Deduplicate some Inlay definitions
- Remove match conversion for InlayKind since we're using remote
